Refactor pipeline and application code for improved clarity; remove unused variables, enhance cleanup processes, and ensure consistent type usage in lambda functions

+++ src/Render/DX12/PipelineBase.cpp
@@ -126,7 +126,7 @@ void PipelineBase::SetGraphicsRoot_CBV_SRV(
     D3D12_SHADER_DESC shaderDesc;
     ThrowIfFailed(refl->GetDesc(&shaderDesc));
 
-    auto GetSRVRootParamIndex = [&](size_t registerIndex) {
+    auto GetSRVRootParamIndex = [&](UINT registerIndex) {
       for (size_t i = 0; i < material->shader->rootParameters.size(); i++) {
         const auto& param = material->shader->rootParameters[i];
 
@@ -146,13 +146,13 @@ void PipelineBase::SetGraphicsRoot_CBV_SRV(
             },
             param);
 
-        if (flag) return i;
+        if (flag) return (UINT)i;
       }
       assert(false);
-      return static_cast<size_t>(-1);
+      return static_cast<UINT>(-1);
     };
 
-    auto GetCBVRootParamIndex = [&](size_t registerIndex) {
+    auto GetCBVRootParamIndex = [&](UINT registerIndex) {
       for (size_t i = 0; i < material->shader->rootParameters.size(); i++) {
         const auto& param = material->shader->rootParameters[i];
 
@@ -170,10 +170,10 @@ void PipelineBase::SetGraphicsRoot_CBV_SRV(
             },
             param);
 
-        if (flag) return i;
+        if (flag) return (UINT)i;
       }
       assert(false);
-      return static_cast<size_t>(-1);
+      return static_cast<UINT>(-1);
     };
 
     for (UINT i = 0; i < shaderDesc.BoundResources; i++) {
@@ -193,7 +193,7 @@ void PipelineBase::SetGraphicsRoot_CBV_SRV(
           if (target == material->properties.end()) break;
 
           auto dim = rsrcDesc.Dimension;
-          size_t rootParamIndex = GetSRVRootParamIndex(rsrcDesc.BindPoint);
+          UINT rootParamIndex = GetSRVRootParamIndex(rsrcDesc.BindPoint);
           switch (dim) {
             case D3D_SRV_DIMENSION_TEXTURE2D: {
               assert(std::holds_alternative<const Texture2D*>(target->second));
